ffi: Use cbindgen to generate a C header file.
Add a build script and use it to generate a description of the
ffi bindings as part of `cargo build` based on the configuration
in cbindgen.toml.

Currently cbindgen.toml just specifies C instead of the default
C++ output, which could be accomplished by adding .with_language()
to the builder configuration. However, the external config also
works with cbindgen is invoked from the command line, and is easier
to read if we add more customization later.