ffi: Add a simple linkage test for the ffi bindings.
Add a C test harness and a GNU Makefile for building it and linking
against the ffi crate's staticlib output. Invoking `make check`
should give some indication whether binding generation is working.

Note that the location of libffi.a is under cargo's control. The
hard-coded path in the Makefile will be brittle.