ppoprf: Merge declaration and assignment.
Address a clippy needless_late_init warning by using the
conditional directly as an initializer. This version elides
the type from the declaration, which I think is fine given
the definition is immediately above.

This isn't failing CI because the warning isn't reported by
the clippy from the MSRV 1.55, but it is reported by clippy
0.1.59 from current stable rust 1.59.0.