ffi: Mark functions unsafe.
Mark functions dereferencing a raw pointer as unsafe to silence
the not_unsafe_ptr_arg_deref clippy lint.

I've often wondered about this. Limiting `unsafe` decorations
to interior blocks reduces the amount of code which must be
reviewed for safety. On the other hand, it's often impossible
when interfacing with foreign functions to guarantee Rust's
definition of safety, so the "safe if used properly" contract
represented by safe function declarations is misleading.

Since clippy has an opinion, go with that. These functions are
small wrappers, so the additional review burden is small. This
style does imply moving any more substantial code into safe
helper functions.

The `unsafe` annotation isn't part of the C ffi bindings, so
there is no change on that side.